Fix .env search path
When `REVIEWS_PATH_TO_CONFIG` is not set, `decouple.AutoConfig()` would
search  for a `.env` file starting from the location of the reviews Python
source, which in most cases was not the appropriate search semantics.

By having `AutoConfig` start in the current directory (`'.'`), we will get
the behavior that most people probably expect.

Closes: #580